Make error messages readable on non-hdpi screens
The error messages are very difficult to read due to the small font, low contrast between the font color and bg color, and the font smoothing. While it is okay when looking at it on e.g. a MacBook it becomes very hard to read on an external monitor with a bit lower pixel density.

Just using the editors font and and slightly increasing the font size makes it perfect.